+++ cosmpy/clients/ledger.py
@@ -26,7 +26,7 @@ import json
 import re
 import time
 from pathlib import Path
-from typing import Callable, List, Optional, Pattern, Tuple, Type, Union
+from typing import List, Optional, Pattern, Tuple, Union
 
 import certifi
 import grpc
@@ -41,6 +41,7 @@ from cosmpy.bank.rest_client import BankRestClient
 from cosmpy.clients.crypto import CosmosCrypto
 from cosmpy.common.loggers import get_logger
 from cosmpy.common.rest_client import RestClient
+from cosmpy.common.retrier import Retrier
 from cosmpy.common.types import JSONLike
 from cosmpy.cosmwasm.rest_client import CosmWasmRestClient
 from cosmpy.protos.cosmos.auth.v1beta1.auth_pb2 import BaseAccount
@@ -109,79 +110,6 @@ class LedgerServerNotAvailable(Exception):
     """
     Ledger server is not available by address provided.
     """
-
-
-class Retrier:
-    def __init__(
-        self,
-        n_retries: int,
-        retry_interval: float,
-        log_retries: bool = True,
-        call_name: str = "execution",
-        exception_type: Type[Exception] = BroadcastException,
-    ):
-        """
-        Call retrier
-
-        :param n_retries: Number of retry attempts
-        :param retry_interval: Length of sleep between retries
-        :param log_retries: bool if retries should be logged
-        :param call_name: str call name for logging
-        :param exception_type: Type of exception to be raised when call fails to execute after multiple attempts
-        """
-
-        self.n_retries = n_retries
-        self.retry_interval = retry_interval
-        self.log_retries = log_retries
-        self.call_name = call_name
-        self.exception_type = exception_type
-
-    def call_with_retry(
-        self,
-        call: Callable,
-        *args,
-        **kwargs,
-    ):
-        """
-        Try to execute call, retry if exception is thrown
-
-        :param call: Callable to be called with args and kwargs
-        :param args: Args to be passed to call
-        :param kwargs: Kwargs to be passed to call
-
-        :raises exception_type: When retry fails after specified number of attempts
-
-        :return: response returned from call
-        """
-
-        last_exception = None
-        response = None
-
-        attempt = 0
-        while attempt < self.n_retries:
-            attempt += 1
-            try:
-                response = call(*args, **kwargs)
-                if response is not None:
-                    break
-            except Exception as e:  # pylint: disable=W0703
-                last_exception = e
-                if self.log_retries:
-                    _logger.warning(
-                        "%s failed, retry in %s seconds: %s",
-                        self.call_name,
-                        self.retry_interval,
-                        e,
-                    )
-                time.sleep(self.retry_interval)
-                continue
-
-        if response is None:
-            raise self.exception_type(
-                f"{self.call_name} failed after multiple attempts: {last_exception}"
-            ) from last_exception
-
-        return response
 
 
 # Class that provides interface to communicate with CosmWasm/Fetch blockchain
@@ -976,6 +904,7 @@ class CosmosLedger:
             n_retries=retries,
             retry_interval=self.msg_retry_interval,
             call_name="Transaction broadcasting",
+            raise_exception_type=BroadcastException,
         ).call_with_retry(
             self.tx_client.BroadcastTx,
             request=broad_tx_req,
@@ -1006,6 +935,7 @@ class CosmosLedger:
             retry_interval=self.get_response_retry_interval,
             call_name="Getting tx response",
             log_retries=False,
+            raise_exception_type=BroadcastException,
         ).call_with_retry(
             self.tx_client.GetTx,
             request=tx_request,
